4. Added salary analytics

+++ EduNEXT.Application/Dtos/Analytics/SalaryAnalyticsDto.cs
@@ -2,9 +2,9 @@ namespace EduNEXT.Application.Dtos.Analytics;
 
 public class SalaryAnalyticsDto
 {
-    public double ActualSalary { get; set; }
+    public decimal ActualSalary { get; set; }
     
-    public double ExpectedSalary { get; set; }
+    public decimal ExpectedSalary { get; set; }
 
-    public List<double> DaysSalary { get; set; } = [];
+    public List<decimal> DaysSalary { get; set; } = [];
 }
